Split comment classes
