+++ src/Moonglade.Model/CommentListItem.cs
@@ -15,19 +15,4 @@ namespace Moonglade.Model
         public DateTime CreateOnUtc { get; set; }
         public IReadOnlyList<CommentReplyItem> CommentReplies { get; set; }
     }
-
-    public class PostCommentListItem
-    {
-        public string Username { get; set; }
-        public DateTime CreateOnUtc { get; set; }
-        public string CommentContent { get; set; }
-        public IReadOnlyList<CommentReplyItem> CommentReplies { get; set; }
-    }
-
-    public class CommentReplyItem
-    {
-        public DateTime ReplyTimeUtc { get; set; }
-
-        public string ReplyContent { get; set; }
-    }
 }
