+++ src/Moonglade.Model/CommentReplyItem.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace Moonglade.Model
+{
+    public class CommentReplyItem
+    {
+        public DateTime ReplyTimeUtc { get; set; }
+
+        public string ReplyContent { get; set; }
+    }
+}
